agents/ClaudeCode/requirements.txt
# Claude Code agent dependencies
claude-agent-sdk>=0.1.19
pyyaml>=6.0.0
